Show best matching email/cardholder name in sender search results
A sender can have more than one email address and more than one
cardholder name.

This adds a new template tag which makes it possible to implement
A succinct '<best matching item> and more'.

As we want to highlight the search term on the results page,
the template tag extract the best match in the list.