Refactor GA events for form submitted
Before:
The event category for submitting a form was "SecurityForms"
and one event per filter used was sent.

After:
We send one event per filter with the new event category
"form-submitted-single-filter" and an extra one which combines
all filters used in the form as a comma separated string
with the event category "form-submitted-combined-filters".