Show errors when conflicting search filters are used
If the user selects a payment method that is not 'debit card',
some of the sender fields (email, postcode, ip_address) do not
make sense.

Before: the form would validate and the search would return 0
records.

After: the form errors so that the user is aware that those fields
cannot be used when payment method is not 'debit card'.