Only return pending checks related to credits in initial state
This adds the ability to only include checks related to credits
in initial state so that if payments time out, the linked
checks don't show up.

It's behind a feature flag as we are currently running checks
in simulation mode so if we enable this, none of the checks will
appear.