Add Job info form for collecting user research data
Jira: https://dsdmoj.atlassian.net/browse/MTP-1266

This commit adds a form to the Intelligence tool which will collect
information about an user's job title, the area of the prison estate
they work in and what their job title involves.

By using the flags used elsewhere in this app we only redirect a user
to a form if they have not previously completed the information.

We considered setting the user flag in the API but that would mean the
flag text had to be set in 2 places and given other flags are set in
this app when user actions occur (choosing prisons for instance) we
decided that this was the right choice.

The custom validation also ensures that if the user selects 'Other' form
the job title list, they still have to enter a job title to be able to
submit the form.

Co-authored-by: George Schena <george@madetech.com>