Refactor can_skip_confirming_prisons method
This refactors the flag getting mechanism for
`can_skip_confirming_prisons` to use the more verbose way to provide a
default value for the `user.user_data.get('flags')` method.

We could pass in the default directly to `user.user_data.get('flags')`
by add in a default value, `user.user_data.get('flags', [])` but this
isn't how the method is used elsewhere in the codebase and matches
other ways to set the flags to a default empty array, and consistency
is important.