Move Credit Accept or Reject Action to Single Page
Jira: https://dsdmoj.atlassian.net/browse/MTP-1354

This commit moves the accept or reject action used by the FIU team to a
single 'Review' page where both accept or reject actions can be
completed.

This also allows a reason for accepting a credit to be added. The name
of the field `rejection_reason` doesn't reflect that yet, but as changes
to the API are required to not break anything this will follow in a
later commit.

The `fiu_action` name attribute on the accept and reject buttons are the
form values that control the type of action the back end will handle,
if they are removed the accept or reject actions won't work so not too
worried about this unconventional (but prefectly valid) form technique,
and it keeps the form simple as only the value on the button that is
clicked is passed to the Django form.

The AcceptCheck and RejectCheck can also be consolidated into a single
class, since it makes no sense for a single small class to rely on
another small class, where the class is not inherited by any other
class.

With a single new template, we can also remove the `reject_check.html`
template.