Update accept/reject template to match designs
Jira: https://dsdmoj.atlassian.net/browse/MTP-1354

This updates the design for the accept/reject check template to
change the credit details to match the designs in the prototype[1].

The inclusion of the warning that a credit needs attention is the most
complex part of this change, using the same logic as the credit list
page. By moving the `get_need_attention_date` to utlities meant that it
would only be defined once and used in the classes. Utility was decided
on because it has not dependencies at all and means the set time
difference is only defined in one place. A possible improvement might be
passing in what time difference you need - but until it changes, this
keeps the 3 days setting in a single place.

1. https://pm-intelligence-tool.herokuapp.com/delayed_capture_history/review_02