Change positioning method for secondary buttons
On the history/credits to action (formerly pending) check page for FIU
users, the success message after approving or rejectign a credit was
appearing underneath the "Show history" button. This was not ideal and
while it didn't cause any significant problems, swapping to
`display: flex` for layout seemed a no brainer.