MTP-1452 Add assign functionality for security checks
Jira: https://dsdmoj.atlassian.net/browse/MTP-1452

This commit adds the ability to assign a check to a user so that they
can manage the checks they are or want to work on.

As there is no interface and it's initially just going to be done on the
Resolve credit page, we handle the form using a view that will just
redirect back to the resolve_check view.

The handling of GET requests is to get round an issue where django will
redirect a user after logging in to the assign view page with a GET
request if they click the add/remove check button, but their session has
expired.

Co-authored-by: staberinde <toby.petty@digital.justice.gov.uk>