MTP-1594 Add policy change warning to intel tool
Jira: https://dsdmoj.atlassian.net/browse/MTP-1594

To inform staff about the changes in policy coming up, this commit adds
an information box to the dashboard so users know how changes affect
them.

The slightly odd `no_spaced_header` value is to control some weird
spacing that happens beneath the notifications banner, however as it
should be temporary didn't see the value in refactoring the spacing to
work better, and this workaround can be removed when banners are removed

The notifications box is also hidden on search results pages as I didn't
want it getting in the way on pages where people are actually doing
work.