MTP-1598 Add post Nov 2 policy change information
Jira: https://dsdmoj.atlassian.net/browse/MTP-1598

This commit adds the conditional display of post November 2 policy
to the Intel tool.

It uses the NOVEMBER_SECOND_CHANGES_LIVE env var as the conditional as
this site has mention of prisoner capping, which is not related to
bank transfers (slightly different implementation to cashbook app)
changes but should be understandable.