MTP-1837: FAQ (when signed out)
When user is not signed in, the links to "Get help" no longer take them
to the "Submit feedback" page but instead they take them to an FAQ page
with questions related to login problems etc.

This FAQ page still has links to contact us.

⚠️ **NOTE**: Currently the 'Request an account' functionality is not
present in `noms-ops` (see 'Request an account - intel tool' epic below)
so in the interest of making this releasable the question with the link
to 'request an account' take the user to the 'Submit a ticket' page
with a pre-populated message.

Ticket: https://dsdmoj.atlassian.net/browse/MTP-1837
'Request an account - intel tool' epic: https://dsdmoj.atlassian.net/browse/MTP-1812