convert_date_fields() only parse datetimes
All the fields in `fields` are datetimes, so we shouldn't
need to attempt to parse values returned by the API
as 'dates'.

The loop through the parsers has been removed which
means the `continue`/`break` had to be removed as they
now refer to the outer loop (the one through the field
names) - we don't want to break/continue as that would
skip the following fields.

NOTE: [Django 4.x changed the way `parse_datetime()`](https://code.djangoproject.com/ticket/35097)
behaves as it nows parses 'YYYY-MM-DD' strings as
datetimes at midnight, although it shoudln't matter anymore
as the fields are all datetimes and all the parsed strings
in the tests are now datetimes.