Revert "convert_date_fields() only parse datetimes"
This reverts commit 236260688fd9a1697625f02dd2f9b2803653a1aa.

NOTE: Resolve conflict resolved manually to account for changes in imports
in other commits after the reverted commit. I'm going to try and retain
the existing behaviour, e.g. don't change the tests, by trying to detect
if the string being parsed is a datetime (contains 'T') or a date (if not).