Retain `convert_date_fields()`'s old behaviour
[Django 4.x changed the way `parse_datetime()`](https://code.djangoproject.com/ticket/35097)
behaves as it nows parses 'YYYY-MM-DD' strings as
datetimes at midnight. It shouldn't matter in theory
as AFAICT all the fields being converted are actually
datetimes (not dates)...

...However I've updated `convert_date_fields()` to
retain the existing behaviour (e.g. no tests changes).
When converting a field `convert_date_fields()` checks
if the string contains a 'T', and if that's the case
it tries to parse it as a `datetime`, otherwise
it parses it as a `date` (old behaviour).

This was suggested in the PR comment and it is potentially
less risky than changing the way this utility work: https://github.com/ministryofjustice/money-to-prisoners-noms-ops/pull/617/commits/236260688fd9a1697625f02dd2f9b2803653a1aa#r2007749702